rename createResourceAtom to createAsyncAtom
