+++ src/utils/atoms/createResourceAtom/createResourceAtom.test.ts
@@ -1,6 +1,6 @@
 import { expect, test, describe, vi } from 'vitest';
 import { createBooleanAtom, createPrimitiveAtom } from '../createPrimitives';
-import { createResourceAtom } from './createResourceAtom.old';
+import { createResourceAtom } from './createResourceAtom';
 
 const wait = (sec: number) => new Promise((res) => setTimeout(res, sec * 1000));
 const waitCallbackBeenCalled = (fn) => {
