+++ src/utils/atoms/createResourceAtom/index.ts
@@ -1,3 +1,2 @@
-export { createResourceAtom } from './createResourceAtom.old';
-export { createResourceAtom as createResourceAtomExperimental } from './createResourceAtom';
-export type { ResourceAtomType, ResourceAtom } from './createResourceAtom.old';
+export { createResourceAtom } from './createResourceAtom';
+export type { ResourceAtomType, ResourceAtom } from './createResourceAtom';
