+++ src/utils/test/id.ts
@@ -1,0 +1,8 @@
+export const incrementId = (() => {
+  const idCreator = function* () {
+    let i = 0;
+    while (true) yield i++;
+  };
+  const idsGenerator = idCreator();
+  return () => String(idsGenerator.next().value);
+})();
