+++ src/utils/test/index.ts
@@ -1,0 +1,2 @@
+export * from './id';
+export * from './wait';
