+++ src/utils/test/wait.ts
@@ -1,0 +1,7 @@
+export const wait = (sec = 1, opt: { failWithMessage?: string } = {}) =>
+  new Promise((res, rej) =>
+    setTimeout(
+      opt?.failWithMessage ? () => rej({ message: opt.failWithMessage }) : res,
+      sec,
+    ),
+  );
